+++ lib/e_menu/src/baltip/pkgIndex.tcl
@@ -1,4 +1,4 @@
-package ifneeded baltip 1.4.1 [list source [file join $dir baltip.tcl]]
+package ifneeded baltip 1.5 [list source [file join $dir baltip.tcl]]
 
 namespace eval ::baltip {
   variable _ruff_preamble {
@@ -149,9 +149,17 @@ To update a tip's text and options:
 
       ::baltip::update widgetpath text ?options?
 
+To show a tip for a widget that has no "normal" (on hovering) tip, still needs a tip (e.g. on clicking):
+
+      ::baltip::showTip path text ?options?
+
 When you click on a widget with its tip being displayed, the tip is hidden. It is the default behavior of *baltip*, but sometimes you need to re-display the hidden tip. If the widget is a button, you can include the following command in `-command` of the button:
 
       ::baltip::repaint widgetpath
+
+Thi
+
+## Some special tips
 
 The "text" for *listbox* can contain %i wildcard - and in such cases the text means a callback receiving a current index of item to tip:
 
@@ -214,6 +222,10 @@ For example:
       set geom "+([expr {$w+$x}]-W-4)+$y"
       set text "The balloon at the right edge of the window"
       ::baltip tip .win $text -geometry $geom -pause 2000 -fade 2000
+
+To show a balloon under the mouse pointer, e.g. on clicking, timeout, processing etc., the following call is used:
+
+      ::baltip::showBalloon text ?options?
 
 
 ## Command
@@ -326,7 +338,7 @@ Also, you can test *baltip* with *test2_pave.tcl* of [apave package](https://chi
 
 The *baltip* package has been developed with help of these kind people:
 
-  * [Nicolas Bats](https://github.com/sl1200mk2) prompted to add canvas tags' tips and tested *baltip* in MacOS
+  * [Nicolas Bats](https://github.com/sl1200mk2) prompted to add canvas tags' tips, baltip::show procedure and tested *baltip* in MacOS
 
   * [Csaba Nemethi](https://www.nemethi.de/) sent several bug fixes and advices, especially on listbox, treeview and menu tips